* release/5.0.0: (76 commits)
  ZEN-16125 - Backoff improvements.
  Fix the tests
  Set more appropriate defaults
  Add exponential backoff code
  Remove unused field; up a log level
  Remove now-unused parameter
  Save a little on GC by reusing byte buffers
  Refactor websocket connection pool into something manageable
  New License file
  default to port 8080 and convert interface to string using sprintf
  added api version byte
  added unit test for MarshalBinary
  added mergo test code
  updated output_test to use "json" encoding
  fixed shipper options; defaults, overridden by config, overridden by cli
  add back gorilla/websocket
  update godeps to bring in snappy
  Snappy\!
  Implement binary encoding
  Go to signed ints
  ...

+++ simulate/model.go
@@ -6,7 +6,7 @@ type Control struct {
 }
 
 type Metric struct {
-	Timestamp float64                `json:"timestamp"`
+	Timestamp int                    `json:"timestamp"`
 	Metric    string                 `json:"metric"`
 	Value     float64                `json:"value"`
 	Tags      map[string]interface{} `json:"tags"`
